Extract Repository and Datyasource from login to domain layer. Remove userdao from repository and add datasource as resource of data.
