+++ login/src/main/java/com/mito/login/data/UserDataRepositoryImpl.kt
@@ -1,11 +1,9 @@
 package com.mito.login.data
 
+import com.mito.login.domain.UserDataRepository
+import com.mito.login.domain.UserDataSource
 import com.mito.network.dummy_login.domain.User
 import javax.inject.Inject
-
-interface UserDataRepository {
-    suspend fun registerUser(user: User)
-}
 
 class UserDataRepositoryImpl @Inject constructor(private val dataSource: UserDataSource) :
     UserDataRepository {
