+++ login/src/main/java/com/mito/login/domain/DummyLoginDataSource.kt
@@ -1,0 +1,8 @@
+package com.mito.login.domain
+
+import com.mito.network.dummy_login.data.response.LoginResponse
+
+
+interface DummyLoginDataSource {
+    suspend fun login(username: String, password: String): Result<LoginResponse>
+}
