+++ login/src/main/java/com/mito/login/domain/DummyLoginUseCase.kt
@@ -1,7 +1,6 @@
 package com.mito.login.domain
 
 import com.mito.common.usecase.Result
-import com.mito.login.data.DummyLoginRepository
 import com.mito.login.ui.LoginUIModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
