+++ login/src/main/java/com/mito/login/domain/RegisterUserUseCase.kt
@@ -1,6 +1,5 @@
 package com.mito.login.domain
 
-import com.mito.login.data.UserDataRepository
 import com.mito.network.dummy_login.domain.User
 import javax.inject.Inject
