+++ login/src/main/java/com/mito/login/domain/UserDataRepository.kt
@@ -1,0 +1,7 @@
+package com.mito.login.domain
+
+import com.mito.network.dummy_login.domain.User
+
+interface UserDataRepository {
+    suspend fun registerUser(user: User)
+}
